requirements.txt
fastapi
uvicorn
pydantic
PyPDF2
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
